// database.types.ts
export type Database = {
  public: {
    Tables: {
      lessons: {
        Row: {
          auditorium: string | null
          auditorium_amount: number | null
          begin: string | null
          building: string | null
          created_at: string | null
          discipline: string | null
          discipline_oid: number | null
          end: string | null
          kind_of_work: string | null
          lecturer: string | null
          lesson_number_start: number | null
          lesson_oid: number
          modified_at: string | null
          note: string | null
          url1: string | null
          url2: string | null
        }
        Insert: {
          auditorium?: string | null
          auditorium_amount?: number | null
          begin?: string | null
          building?: string | null
          created_at?: string | null
          discipline?: string | null
          discipline_oid?: number | null
          end?: string | null
          kind_of_work?: string | null
          lecturer?: string | null
          lesson_number_start?: number | null
          lesson_oid: number
          modified_at?: string | null
          note?: string | null
          url1?: string | null
          url2?: string | null
        }
        Update: {
          auditorium?: string | null
          auditorium_amount?: number | null
          begin?: string | null
          building?: string | null
          created_at?: string | null
          discipline?: string | null
          discipline_oid?: number | null
          end?: string | null
          kind_of_work?: string | null
          lecturer?: string | null
          lesson_number_start?: number | null
          lesson_oid?: number
          modified_at?: string | null
          note?: string | null
          url1?: string | null
          url2?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof PublicSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof PublicSchema["CompositeTypes"]
    ? PublicSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never